+++ WinFormsUI/WinFormsUI.csproj
@@ -39,18 +39,6 @@
       <DependentUpon>MainForm.cs</DependentUpon>
     </Compile>
     <Compile Include="Program.cs" />
-    <Compile Include="TimeSeriesForm.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="TimeSeriesForm.Designer.cs">
-      <DependentUpon>TimeSeriesForm.cs</DependentUpon>
-    </Compile>
-    <Compile Include="TransactionsForm.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="TransactionsForm.Designer.cs">
-      <DependentUpon>TransactionsForm.cs</DependentUpon>
-    </Compile>
   </ItemGroup>
   <ItemGroup>
     <Reference Include="Microsoft.Win32.Primitives, Version=4.0.2.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a, processorArchitecture=MSIL">
@@ -223,12 +211,6 @@
   <ItemGroup>
     <EmbeddedResource Include="MainForm.resx">
       <DependentUpon>MainForm.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="TimeSeriesForm.resx">
-      <DependentUpon>TimeSeriesForm.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="TransactionsForm.resx">
-      <DependentUpon>TransactionsForm.cs</DependentUpon>
     </EmbeddedResource>
   </ItemGroup>
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
